chatbot/requirements.txt
flask
flask-cors
torch
nltk
scikit-learn
requests
llama-index
mysql-connector-python
